requirements.txt
# Core Airflow
apache-airflow==2.10.2

# Data handling
pandas==2.2.2
numpy==1.26.4

# Machine Learning
scikit-learn==1.5.1
xgboost==2.1.1

# Visualization (for EDA notebook)
matplotlib==3.9.2
seaborn==0.13.2

# Jupyter (optional, for notebooks)
jupyterlab==4.2.5

